requirements.txt
aws-cdk.aws-appstream==1.51.0
aws-cdk.aws-cloudwatch==1.51.0
aws-cdk.aws-events==1.51.0
aws-cdk.aws-events-targets==1.51.0
aws-cdk.aws-iam==1.51.0
aws-cdk.aws-kms==1.51.0
aws-cdk.aws-lambda==1.51.0
aws-cdk.aws_lambda_event_sources==1.51.0
aws-cdk.aws-s3==1.51.0
aws-cdk.aws-s3-assets==1.51.0
aws-cdk.aws-s3-deployment==1.50.0
aws-cdk.aws-sagemaker==1.51.0
aws-cdk.aws-ssm==1.51.0
aws-cdk.core==1.51.0
aws-cdk.aws-cloudformation==1.51.0
aws-cdk.custom-resources==1.51.0
aws_cdk.aws_logs==1.51.0
boto3==1.14.17